New issue: Find products by name and phone number

+++ src/main/java/com/example/web_inventory/services/SupplierService.java
@@ -5,11 +5,14 @@ import java.util.Optional;
 
 import org.hibernate.ObjectNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.example.web_inventory.dtos.request.SupplierRequestDTO;
+import com.example.web_inventory.entities.ProductEntity;
 import com.example.web_inventory.entities.SupplierEntity;
+import com.example.web_inventory.repositories.ProductRepository;
 import com.example.web_inventory.repositories.SupplierRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -20,6 +23,9 @@ public class SupplierService {
     
     @Autowired
     private SupplierRepository repository;
+
+    @Autowired
+    private ProductRepository productRepository;
 
     public SupplierEntity createSupplier(SupplierRequestDTO dto) {
         SupplierEntity entity = new SupplierEntity(dto);
@@ -32,6 +38,10 @@ public class SupplierService {
 
         return supplier.orElseThrow(() -> new ObjectNotFoundException(
                 "Cliente não encontrado! Id: " + id + "Tipo: " + SupplierEntity.class.getName(), supplier));
+    }
+
+    public Optional<SupplierEntity> findSupplierByNameOrContact(String companyName, String contact) {
+        return repository.findByNameOrContact(companyName, contact);
     }
 
     public List<SupplierEntity> getAllSuppliers() {
@@ -51,10 +61,20 @@ public class SupplierService {
     }
 
     public ResponseEntity<Object> deleteSupplierById(Long id) {
-        return repository.findById(id)
-            .map(delete -> {
-                repository.deleteById(id);
-                return ResponseEntity.noContent().build();
-            }).orElse(ResponseEntity.notFound().build());
+        Optional<SupplierEntity> supplier = repository.findById(id);
+    
+            if (supplier.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+
+        List<ProductEntity> products = productRepository.findBySupplierId(id);
+
+        if (!products.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("Não é possível excluir este fornecedor porque ele tem produtos associados.");
+        }
+
+        repository.deleteById(id);
+        return ResponseEntity.noContent().build();
     }
 }
